🚚 rename: Message, Room 모델 위치 변경
